Просто переносим рендеринг доски фильмов из `main.js` в `page-controller.js'.
